+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/utils/media/MediaViewPane.java
@@ -5,7 +5,6 @@ import javafx.geometry.HPos;
 import javafx.geometry.VPos;
 import javafx.scene.Node;
 import javafx.scene.control.ProgressIndicator;
-import javafx.scene.control.Tooltip;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.Pane;
 import javafx.scene.media.MediaView;
@@ -15,7 +14,6 @@ import javafx.scene.media.MediaView;
  * Also manage rotated images or videos
  */
 public class MediaViewPane extends Pane {
-
   private int marginX = 0;
   private int marginY = 0;
 
@@ -24,8 +22,16 @@ public class MediaViewPane extends Pane {
    */
   private Node child;
 
-  /** Whether Image or Video is rotated, When rotated, invert width and height role */
+  /**
+   * Whether Image or Video is rotated, When rotated, invert width and height role
+   */
   private boolean rotated;
+
+  protected MediaOptions mediaOptions;
+
+  public void setMediaOptions(MediaOptions mediaOptions) {
+    this.mediaOptions = mediaOptions;
+  }
 
   /**
    * Set the child in the middle of the Pane
@@ -64,7 +70,7 @@ public class MediaViewPane extends Pane {
     }
     else {
       Bounds bounds = child.getLayoutBounds();
-      child.relocate((width - bounds.getWidth()) / 2.0, (height  - bounds.getHeight()) / 2);
+      child.relocate((width - bounds.getWidth()) / 2.0, (height - bounds.getHeight()) / 2);
     }
   }
 
@@ -76,7 +82,7 @@ public class MediaViewPane extends Pane {
     return rotated;
   }
 
-  protected void setRotated(boolean rotated) {
+  public void setRotated(boolean rotated) {
     this.rotated = rotated;
   }