+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/MediaPreviewController.java
@@ -4,6 +4,7 @@ import de.mephisto.vpin.commons.fx.DialogController;
 import de.mephisto.vpin.commons.utils.WidgetFactory;
 import de.mephisto.vpin.commons.utils.media.AssetMediaPlayer;
 import de.mephisto.vpin.commons.utils.media.ImageViewer;
+import de.mephisto.vpin.commons.utils.media.MediaOptions;
 import de.mephisto.vpin.restclient.frontend.VPinScreen;
 import de.mephisto.vpin.restclient.games.FrontendMediaItemRepresentation;
 import de.mephisto.vpin.restclient.games.GameRepresentation;
@@ -27,6 +28,7 @@ import static de.mephisto.vpin.ui.Studio.client;
 
 public class MediaPreviewController implements Initializable, DialogController, ChangeListener<Number> {
   private final static Logger LOG = LoggerFactory.getLogger(MediaPreviewController.class);
+  public static final int MARGIN = 44;
 
 
   @FXML
@@ -49,16 +51,19 @@ public class MediaPreviewController implements Initializable, DialogController, 
 
   public void setData(Stage dialogStage, GameRepresentation game, FrontendMediaItemRepresentation item) {
     this.dialogStage = dialogStage;
-    assetMediaPlayer = WidgetFactory.addMediaItemToBorderPane(client, item, mediaView);
+
+    MediaOptions mediaOptions = new MediaOptions();
+    mediaOptions.setAutoRotate(false);
+    assetMediaPlayer = WidgetFactory.addMediaItemToBorderPane(client, item, mediaView, null, mediaOptions);
     this.item = item;
     Platform.runLater(() -> {
       if (assetMediaPlayer == null) {
         ImageViewer imageViewer = (ImageViewer) mediaView.getUserData();
-        imageViewer.getImageView().setFitWidth(dialogStage.getWidth() * 1 - 80);
-        imageViewer.getImageView().setFitHeight(dialogStage.getHeight() * 1 - 80);
+        imageViewer.getImageView().setFitWidth(dialogStage.getWidth() * 1 - MARGIN);
+        imageViewer.getImageView().setFitHeight(dialogStage.getHeight() * 1 - MARGIN);
       }
       else {
-        assetMediaPlayer.setMediaViewSize(dialogStage.getWidth() * 1 - 80, dialogStage.getHeight() * 1 - 80);
+        assetMediaPlayer.setMediaViewSize(dialogStage.getWidth() * 1 - MARGIN, dialogStage.getHeight() * 1 - MARGIN);
       }
     });
 
@@ -75,13 +80,13 @@ public class MediaPreviewController implements Initializable, DialogController, 
   public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
     if (assetMediaPlayer == null) {
       ImageViewer imageViewer = (ImageViewer) mediaView.getUserData();
-      imageViewer.getImageView().setFitWidth(dialogStage.getWidth() * 1 - 80);
-      imageViewer.getImageView().setFitHeight(dialogStage.getHeight() * 1 - 80);
+      imageViewer.getImageView().setFitWidth(dialogStage.getWidth() * 1 - MARGIN);
+      imageViewer.getImageView().setFitHeight(dialogStage.getHeight() * 1 - MARGIN);
     }
     else {
       VPinScreen screen = VPinScreen.valueOf(item.getScreen());
       boolean rotated = VPinScreen.PlayField.equals(screen) || VPinScreen.Loading.equals(screen);
-      assetMediaPlayer.setMediaViewSize(dialogStage.getWidth() * 1 - 80, dialogStage.getHeight() * 1 - 80);
+      assetMediaPlayer.setMediaViewSize(dialogStage.getWidth() * 1 - MARGIN, dialogStage.getHeight() * 1 - MARGIN);
     }
   }
 }